Cleanup redundant InstanceShardOperationRequest
As stated by the TODO, this thing can just be inlined into the update
request which I have done here.
